refactor(task): keep roster rendering test-local
